+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ CHANGELOG
 * New class method `prep_index_chunk` on ``Indexable`` class, to support
   prefetching related objects when iterating over Django querysets for indexing
 * Include django view mixins in sphinx documentation  
+* Dropped support for python 3.6; added python 3.9
+* Dropped support for Django 2.2; added Django 3.2
 
 0.8.2
 -----